Added the option "note?mac=11:22:11:22:11:22&note=Hello+World To store a note to a mac on the database. The new table is Details. The columns so far are: Mac (text primary key) and Note (text)
I'm performing a STRONG verification on the mac address and the note.
I'm replacing the '+' strings in the note to ' '. So you can send a Text
like this "Hello World", and the web browser can convert it to
"Hello+World".

If the mac does not have 6 pairs of values, if the length of the mac is not
17, if the mac has any character besides "a-zA-Z0-9:"...
OR
if the note has any character besides "a-zA-Z0-9 " ..... then
nothing happens and the web server answers a blank page.

If the insertion is successful, then the web server answers Ok.